+++ tests/sqlparser_common.rs
@@ -25,9 +25,31 @@ use sqlparser::ast::*;
 use sqlparser::dialect::{GenericDialect, PostgreSqlDialect, SQLiteDialect};
 use sqlparser::keywords::ALL_KEYWORDS;
 use sqlparser::parser::{Parser, ParserError};
+use sqlparser::tokenizer::Token;
 use test_utils::{
     all_dialects, expr_from_projection, join, number, only, table, table_alias, TestedDialects,
 };
+
+#[test]
+fn test_prev_index() {
+    let sql = "SELECT version";
+    all_dialects().run_parser_method(sql, |parser| {
+        assert_eq!(parser.peek_token(), Token::make_keyword("SELECT"));
+        assert_eq!(parser.next_token(), Token::make_keyword("SELECT"));
+        parser.prev_token();
+        assert_eq!(parser.next_token(), Token::make_keyword("SELECT"));
+        assert_eq!(parser.next_token(), Token::make_word("version", None));
+        parser.prev_token();
+        assert_eq!(parser.peek_token(), Token::make_word("version", None));
+        assert_eq!(parser.next_token(), Token::make_word("version", None));
+        assert_eq!(parser.peek_token(), Token::EOF);
+        parser.prev_token();
+        assert_eq!(parser.next_token(), Token::make_word("version", None));
+        assert_eq!(parser.next_token(), Token::EOF);
+        assert_eq!(parser.next_token(), Token::EOF);
+        parser.prev_token();
+    });
+}
 
 #[test]
 fn parse_insert_values() {
