+++ tests/sqlparser_hive.rs
@@ -15,8 +15,10 @@
 //! Test SQL syntax specific to Hive. The parser based on the generic dialect
 //! is also tested (on the inputs it can handle).
 
+mod test_utils;
+
 use sqlparser::dialect::HiveDialect;
-use sqlparser::test_utils::*;
+use test_utils::*;
 
 #[test]
 fn parse_table_create() {
